renaming motion_detect file to motion
